% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{balanzasLPAQ}
\alias{balanzasLPAQ}
\title{Especificaciones para balanzas del LPAQ}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 7 columns.
}
\usage{
data(balanzasLPAQ)
}
\description{
Tolerancias especificadas por el fabricantes para balanza ME204 (Mettler Toledo), AL54 (Mettler Toledo) y AES220-4 (Kern)
}
\examples{
data(balanzasLPAQ)
}
\keyword{datasets}
