% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{condAmbientales}
\alias{condAmbientales}
\title{Condiciones ambientales}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(condAmbientales)
}
\description{
Datos de temperatura, Humedad relativa (HR) y presión en un laboratorio sin control climático en Bogotá
}
\examples{
data(condAmbientales)
}
\keyword{datasets}
