% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datosAnalisis}
\alias{datosAnalisis}
\title{Resultados de analisis}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(datosAnalisis)
}
\description{
Conjunto de datos resultado de analisis químico de diferentes tipos de muestras
}
\examples{
data(datosAnalisis)
}
\keyword{datasets}
