% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{densidad_aire}
\alias{densidad_aire}
\title{Densidad del aire}
\usage{
densidad_aire(T_, p, HR, x_CO2 = 4e-04)
}
\arguments{
\item{T_}{temperatura del aire en ºC}

\item{p}{presión expresada en hPa}

\item{HR}{porcentaje de humedad relativa}

\item{x_CO2}{fracción molar de CO2 (por omisión 0.00040)}
}
\value{
densidad del aire a las condiciones ambientales especificadas en kg/m^3
}
\description{
Devuelve la densidad del aire a la condiciones ambientales especificadas, empleando la ecuación
revisada por el CIPM en 2007 (1). Esta ecuación es válida para 600 hPa <= p <= 1100 hPa y 15 ºC <= T <= 27 ºC.
}
\examples{
densidad_aire(T_=20,p=750,HR=50)

}
\references{
(1) Picard, A., Davis, R.S.,  Gläser, M., Fujii, K., Revised formula for the density of moist air (CIPM-2007), Metrologia, 45, 149-155 (2004) \url{https://doi.org/10.1088/0026-1394/45/2/004}.
}
