% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EMP_materialVolumetrico}
\alias{EMP_materialVolumetrico}
\title{Errores máximos permitidos para material volumétrico}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(EMP_materialVolumetrico)
}
\description{
Errores máximos permitidos para buretas (ISO 385:2005), pipetas (ISO 648:2008), balones aforados (ISO 1042:1998) y micropipetas (ISO 8655-2:2002).
}
\examples{
data(EMP_materialVolumetrico)
}
\keyword{datasets}
