% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flotacion}
\alias{flotacion}
\title{Efecto de flotación}
\usage{
flotacion(
  den_s,
  T_ = c(21.5, 15.7, 24.7),
  p = c(749.6, 746, 753.8),
  HR = c(47, 31.2, 67.4),
  den_cal = c(8006, 10),
  den_aire_cal = c(1.2, 0.0129),
  x_CO2 = 4e-04
)
}
\arguments{
\item{den_s}{c(denMuestra,u_denMuestra) densidad e incertidumbre estándar de la muestra que se mide en kg/m^3}

\item{T_}{c(T_prom,T_min,T_max) temperatura del aire en ºC}

\item{p}{c(p_prom,p_min,p_max) presión expresada en hPa}

\item{HR}{c(HR_prom,HR_min,HR_max) porcentaje de humedad relativa}

\item{den_cal}{c(denCal,u_denCal) densidad de las pesas empleadas para la calibración e incertidumbre asociada en kg/m^3,
por omision den_cal=c(8006,10)}

\item{den_aire_cal}{c(denAireCal,u_denAireCal) densidad e incertidumbre estándar del aire durante el ajuste de la sensibilidad de la balanza en kg/m^3,
por omision den_aire_cal=c(1.2,0.0129)}

\item{x_CO2}{fracción molar de CO2 (por omisión 0.00040)}
}
\value{
data.frame con factor de corrección debido al efecto de flotación (Bu), incertidumbre estáncadar (u_Bu)
y densidad (das) e incertidumbre estándar del aire (u_das), en el ambiente del laboratorio.
}
\description{
Devuelve el  factor de corrección debido al efecto de flotación (pesado en aire)
\deqn{Bu=\frac{1-\rho_{\mathrm{aire,cal}}/\rho_{\mathrm{pesas,cal}}}{ 1-\rho_{\mathrm{aire,med}}/\rho_{\mathrm{muestra}}}}
}
\examples{
flotacion(c(998,30),p=c(1010,995,1025),T_=c(22,19,25),HR=c(50,25,75),den_aire_cal=c(1.19,0.0129))

}
\references{
(1) Picard, A., Davis, R.S.,  Gläser, M., Fujii, K., Revised formula for the density of moist air (CIPM-2007), Metrologia, 45, 149-155 (2004) \url{https://doi.org/10.1088/0026-1394/45/2/004}.
}
