% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{masa_minima}
\alias{masa_minima}
\title{Minima cantidad a pesar}
\usage{
masa_minima(
  u_r,
  densidad,
  balanza = "ME204",
  u_densidad = 0.05,
  T_ = c(21.5, 15.7, 24.7),
  p = c(749.6, 746, 753.8),
  HR = c(47, 31.2, 67.4),
  den_cal = c(8006, 10),
  den_aire_cal = c(1.2, 0.0129),
  x_CO2 = 4e-04,
  polyOrden = 4
)
}
\arguments{
\item{u_r}{incertidumbre estándar relativa requerida, expresada porcentualmente.}

\item{densidad}{densidad de la muestra en kg/m^3}

\item{balanza}{empleada:"ME204" (valor por omisión),"AL54" y AS220}

\item{u_densidad}{incertidumbre estándar relativa en la densidad de la muestra, 5 \% por omisión y tomada como una distribución rectangular.
Si se conoce la incertidumbre estándar de la densidad, se debe incluir como u_den*sqrt(3)}

\item{T_}{c(T_prom,T_min,T_max) temperatura del aire en ºC}

\item{p}{c(p_prom,p_min,p_max) presión expresada en hPa}

\item{HR}{c(HR_prom,HR_min,HR_max) porcentaje de humedad relativa}

\item{den_cal}{c(denCal,u_denCal) densidad de las pesas empleadas para la calibración e incertidumbre asociada en kg/m^3,
por omision den_cal=c(8006,10)}

\item{den_aire_cal}{c(denAireCal,u_denAireCal) densidad e incertidumbre estándar del aire durante el ajuste de la sensibilidad de la balanza en kg/m^3,
por omision den_aire_cal=c(1.2,0.0129)}

\item{x_CO2}{fracción molar de CO2 (por omisión 0.00040)}

\item{polyOrden}{grado de polinomio usado para ajustar la curva caracterítica de la balanza}
}
\value{
Masa minima a pesar para la incertidumbre estándar especificada y curva caracterítica de la balanza
}
\description{
Devuelve mínima cantidad a pesar para una incertidumbre estandar de pesada requerida
}
\references{
(1) Reichmuth, A., Wunderli, A., Weber, M., Meyer, V.R., The Uncertainty of Weighing Data Obtained with Electronic Analytical Balances,
Microchim. Acta 148, 133–141 (2004). \url{https://doi.org/10.1007/s00604-004-0278-3}
}
