% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{puntoFinal}
\alias{puntoFinal}
\title{determina el punto final en una curva de titulación}
\usage{
puntoFinal(x_, y_, intervalo_x = c(0, 0))
}
\arguments{
\item{x_}{conjunto de datos x asociados a la curva de titulación}

\item{y_}{conjunto de datos y asociados a la curva de titulación}

\item{intervalo_x}{corresponde al intervalo en el que busca el punto de inflexión de la curva dada}
}
\value{
data.frame con factor con abcisa en el punto de equivalencia e incertidumbre estándar asociada a la determinación
}
\description{
Obtiene el punto final de una curva de titulación en el intervalo expecificado, empleando diferentes enfoques:
i)   Máximo de la primera derivada numérica. La incertidumbre se obtiene asumiendo un distribución triangular entre los dos puntos adyacentes al valor encontrado, dentro del conjunto de datos crudos.
ii)  Punto de inflexión obtenido empleando estimador de distancia extrema del paquete inflection (1)
iii) Punto de inflexión empleando modelo de regresión logística empleando paquete nplr
}
\details{
La incertidumbre devuelta corresponde al resultado de la estimación empleando los algoritmos indicados.
}
\references{
(1) Demetris T. Christopoulos, On the Efficient Identification of an Inflection Point, International Journal of Mathematics and Scientific Computing, 6 (1), 13-20, 2016.
\url{https://veltech.edu.in/wp-content/uploads/2016/04/Paper-04-2016.pdf}.
}
\seealso{
\code{\link[=edeci]{edeci()}},\code{\link[=nplr]{nplr()}}
}
