% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pureza}
\alias{pureza}
\title{Estimado de pureza e incertidumbre basados en la ficha técnica del estándar.}
\usage{
pureza(P, delta = 0, tecnica = "n", sInstrum = FALSE, dist = "rectangular")
}
\arguments{
\item{P}{pureza, pureza minima o vector con pureza minima y pureza máxima c(P_min,P_max), expresadas como fracción másica.}

\item{delta}{variación indicada sin especificación de factor de cobertura o tipo de distribución. Si delta=0, se asume que P es una pureza mínima}

\item{tecnica}{técnica de análisis descrita: "titulacion", "cromatografia", "n" (no indicada)}

\item{sInstrum}{indica si el delta proveido corresponde a la incertidumbre instrumental (FALSE, por omisión)}

\item{dist}{distribucion asumida para casos donde se especifica un estimado de P y delta sin conocimiento de tipo de distribucion o factor de cobertura; o valores máximos y mínimos de P. Por omisión se asume una distribución "rectangular".
Sí solo se indica un valor informativo estimado de P, sin información sobre incertidumbre, use delta=1E-10 y dist="triangular" para obtener un estimado de la incertidumbre estándar con valor centrado en P y distribución triangular en el intervalo +/-(1-P).}
}
\value{
data.frame con pureza e incertidumbre estimadas a partir de la especificaciones
}
\description{
Devuelve estimados de puereza en incertidumbre basados en especificaciones técnicas del reactivo (1-2)
Los valores se deben ingresar como fracciones másicas en el intervalo 0-1. Sí se trata de un MRC no es necesario emplear esta función.
}
\examples{
pureza(P=0.999) # Solo se indica el límite inferior de la pureza

}
\references{
(1) van Look, G., Meyer, V.R., The purity of laboratory chemicals with regard to measurement uncertainty, 	Analyst, 127, 825-829, 2002. \url{https://doi.org/10.1039/B107958C}

(2) Borges, R., Meyer, V.R. The uncertainty of purity of reference materials must be known, J. of Pharm. and Biomed. Anal., 77, 40-43, 2013. \url{https://doi.org/10.1016/j.jpba.2013.01.008}
}
