% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{u_masa}
\alias{u_masa}
\title{Incertidumbre en mediciones de masa, empleando especificaciones del fabricante (balanzas sin calibración)}
\usage{
u_masa(
  lectura,
  densidad,
  balanza = "ME204",
  u_densidad = 0.05,
  T_ = c(21.5, 15.7, 24.7),
  p = c(749.6, 746, 753.8),
  HR = c(47, 31.2, 67.4),
  den_cal = c(8006, 10),
  den_aire_cal = c(1.2, 0.0129),
  x_CO2 = 4e-04
)
}
\arguments{
\item{lectura}{Lectura de masa en unidad "g", obtenida del instrumento de medición.}

\item{densidad}{densidad de la muestra en kg/m^3}

\item{balanza}{empleada:"ME204" (valor por omisión),"AL54" y AS220}

\item{u_densidad}{incertidumbre estándar relativa en la densidad de la muestra, 5 \% por omisión y tomada como una distribución rectangular.
Si se conoce la incertidumbre estándar de la densidad, se debe incluir como u_den*sqrt(3)}

\item{T_}{c(T_prom,T_min,T_max) temperatura del aire en ºC}

\item{p}{c(p_prom,p_min,p_max) presión expresada en hPa}

\item{HR}{c(HR_prom,HR_min,HR_max) porcentaje de humedad relativa}

\item{den_cal}{c(denCal,u_denCal) densidad de las pesas empleadas para la calibración e incertidumbre asociada en kg/m^3,
por omision den_cal=c(8006,10)}

\item{den_aire_cal}{c(denAireCal,u_denAireCal) densidad e incertidumbre estándar del aire durante el ajuste de la sensibilidad de la balanza en kg/m^3,
por omision den_aire_cal=c(1.2,0.0129)}

\item{x_CO2}{fracción molar de CO2 (por omisión 0.00040)}
}
\value{
lista con:
i)   data.frame (ms) para masa corregida por efecto de flotación (ms) e incertidumbre estándar en la medición de masa (u_ms) expresada en "g"
ii)  data.frame (Bu) con corrección por efecto de flotación (adimensional), y factores de influencia considerados: a) densidad de la muestra (d_s), b) densidad del aire durante la medición (d_as), c) densidad del aire durante la ajuste y calibración (d_ac), c) densidad de las pesas empleandas para ajuste y calibración (d_pc)
iii) data.frame (Ws) con incertidumbre debidad al instrumento  (ws) en "mg", y sus factores de influencia: a) Resolución (u_Res), b) repetibilidad (u_rep), c) desvio de la linealidad (u_NL), d) Tolerancia en la sensibilidad (u_TS), e) Efecto térmico en la sensibilidad (u_ETS)
}
\description{
Devuelve la incertidumbre debida a una operación de pesaje en una balanza de la que no se tiene información de
calibración. Teniendo en cuenta las siguientes consideraciones:
i)   La función solo considera las balanzas que se emplean en el LPAQ (ME204, AL54 y AS220)
ii)  La muestra y el recipiente en que se mide, se colococan centrados en el plato (i.e. no se considera el efecto de la excentricidad)
iii) La muestra tiene masa definida i.e. efectos de volatilidad de la muestra o higroscopicidad son despreciables
iv)  El operador de la balanza conoce el procedimiento de pesaje y lo efectua con el debido cuidado
}
\examples{
u_masa(.0010,1636)

}
\references{
(1) Reichmuth, A., Wunderli, A., Weber, M., Meyer, V.R., The Uncertainty of Weighing Data Obtained with Electronic Analytical Balances,
Microchim. Acta 148, 133–141 (2004). \url{https://doi.org/10.1007/s00604-004-0278-3}
}
