% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{u_VT}
\alias{u_VT}
\title{Incertidumbre en el volumen de un líquido debido a variaciones de temperatura}
\usage{
u_VT(volumen, DT, gamma = 0.00021, distrib = "rectangular")
}
\arguments{
\item{volumen}{Volumen utilizado o entregado en la medición}

\item{DT}{temperatura respecto a la temperatura estándar de referencia (normalmente 20 ºC) en grados Celsius}

\item{gamma, }{coefficiente de expansión térmico (por omisión gamma=2.1E-4/ºC para el agua)}

\item{distrib}{distribución asumida (por omisión distrib="rectangular")}
}
\value{
incertidumbre estándar en el volumen del líquido debido a variación de temperatura respecto a la referencia.
La incertidumbre calculada se obtiene asumiendo una distribución rectangular para la variación de temperatura.
}
\description{
Devuelve la incertidumbre en el volumen del líquido empleado, debido a variaciones de temperatura  respecto a la temperatura estándar de referencia de 20 ºC.
No se considera la expansión del material del instrumento, dado que los coeficientes de expansión térmica del vidrio son más bajos que los del líquido (Borosilicato 3.3 9.9E-6/ºC
y vidrio común 27E-6/ºC). El efecto de la temperatura sobre el vidrio produce variaciones en volumen de 0.007 \% para el borosilicato y 0.02 \% para el vidrio común
a una temperatura de 27 ºC, evidenciando un  efecto es mínimo de la temperatura sobre la expansión del material del instrumento (1).
}
\examples{
u_VT(volumen=100.0,DT=4)

}
\references{
(1) ISO 4787:2010, Laboratory glassware - Volumetric instruments - Methods for testing of capacity and for use. Geneva, Switzerland. \url{https://www.iso.org/standard/41807.html}
}
